pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opsfoundry"
version = "0.1.0"
description = "Modular AI operations platform for workflow automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "Pillow>=11.1,<12.0",
    "mlflow>=2.20,<3.0",
    "PyYAML>=6.0,<7.0",
    "pypdfium2>=4.30,<5.0",
    "python-multipart>=0.0.9,<1.0",
    "redis>=5.2,<6.0",
    "SQLAlchemy>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "httpx>=0.28,<1.0",
    "ruff>=0.11,<1.0",
]

[tool.setuptools.packages.find]
include = [
    "apps*",
    "core*",
    "workflows*",
    "mcp_servers*",
]

[tool.setuptools.package-data]
"workflows.invoice_autoposting" = ["config.yaml"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
